pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-react-agent"
version = "1.0.0"
description = "ARIA: Multi-agent AI assistant with web search and document understanding"
authors = [
    {name = "Pamudi", email = "Pamu0002@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "langchain==0.3.7",
    "langchain-groq==0.1.8",
    "langchain-community==0.2.0",
    "fastapi==0.115.4",
    "uvicorn==0.30.0",
    "pydantic==2.8.2",
    "pydantic-settings==2.3.1",
    "requests==2.32.3",
    "aiofiles==23.2.1",
    "python-dotenv==1.0.1",
    "tavily-python==0.3.16",
    "beautifulsoup4==4.12.3",
    "pypdf==4.3.0",
    "python-docx==1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.2",
    "pytest-asyncio==0.23.6",
    "black==24.8.0",
    "isort==5.13.2",
    "flake8==7.1.0",
    "mypy==1.11.2",
]

[project.urls]
Homepage = "https://github.com/Pamu0002/langchain-react-agent"
Repository = "https://github.com/Pamu0002/langchain-react-agent.git"
Documentation = "https://github.com/Pamu0002/langchain-react-agent/blob/main/README.md"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | \.pytest_cache
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
